Creating new reservation with validation for start time of projection

+++ CinemAPI.Data.Implementation/ProjectionRepository.cs
@@ -23,7 +23,7 @@ namespace CinemAPI.Data.Implementation
                                                       x.StartDate == startDate);
         }
 
-        public IProjection GetById(int id)
+        public IProjection GetById(long id)
         {
             return db.Projections.FirstOrDefault(x => x.Id == id);
         }
